added ZombiesOMG.pde to repository
ZombiesOMG is a main program loop. The code was originally authored in
March of 2013 under the name SimpleFigure version 3.  It invokes the
SimpleFigure class to create individuals (Tom, Dick, and Harry), and a
variable population array of zombies. A mouse click on the canvas
should move the zombies to the mouse location and increment the zombie
index.
Known Issues: the code was functional as of March. In June it began to
behave erratically. Cause is unknown. There have been changes to the
Java version and the Processing version has gone from 2.0beta to 2.0.
These may cause or contribute to the problem.